refractor hash functions
